requirements.txt
--extra-index-url https://pip.hexaly.com

cvxpy-base==1.6.0
numpy
scipy==1.15.1
plotly==5.24.1

# we also install them on the fly in the Makefile
marimo==0.10.12
pre-commit==4.0.1

# Install solvers
clarabel==0.9.0
mosek==10.2.13
hexaly

# for testing
pytest
pytest-cov
